# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion des dettes et créances",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Gestion des dettes et récues"
        }
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Authentification d'un utilisateur",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "idrissa.wade"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "SecureP@ss2024!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJhbGci..."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion reussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "411": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 411
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Échec de l'authentification",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Echec de l'authentification"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Créer un compte utilisateur pour un client",
                "description": "Crée un compte utilisateur pour un client existant (rôle Boutiquier requis)",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password",
                                    "clientId"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "client.user"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "SecureP@ss2024!"
                                    },
                                    "clientId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "data": {
                                            "properties": {
                                                "client": {
                                                    "$ref": "#/components/schemas/Client"
                                                },
                                                "user": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur cree avec succees"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "411": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 411
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion de l'utilisateur",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    },
                    "500": {
                        "description": "Erreur lors de la déconnexion"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Client": {
                "required": [
                    "Surname",
                    "Telephone"
                ],
                "properties": {
                    "Surname": {
                        "type": "string",
                        "example": "John"
                    },
                    "Telephone": {
                        "type": "number",
                        "example": "77#######"
                    },
                    "Adresse": {
                        "type": "string",
                        "example": "123 Rue Principale"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "required": [
                    "nom",
                    "prenom",
                    "login",
                    "role"
                ],
                "properties": {
                    "nom": {
                        "type": "string",
                        "example": "John"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "login": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "role": {
                        "type": "string",
                        "example": "CLIENT"
                    },
                    "photo": {
                        "type": "string",
                        "example": "/uploads/users/photo.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}